% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstFunction.R
\name{myFirstFunction}
\alias{myFirstFunction}
\title{Renders x}
\usage{
myFirstFunction(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
The input
}
\description{
Renders x
}
\examples{
myFirstFunction(x)

}
