% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStatisticalMode.R
\name{myStatisticalMode}
\alias{myStatisticalMode}
\title{Statistical Mode}
\usage{
myStatisticalMode(x)
}
\arguments{
\item{x}{a vector containing numeric values}
}
\description{
Statistical Mode
}
\examples{
myStatisticalMode(c(1,2,1,3))
}
